+++ src/_pytask/clean.py
@@ -29,8 +29,8 @@ from _pytask.pluginmanager import get_plugin_manager
 from _pytask.session import Session
 from _pytask.shared import to_list
 from _pytask.traceback import render_exc_info
+from _pytask.tree_util import tree_leaves
 from attrs import define
-from pybaum.tree_util import tree_just_yield
 
 
 if TYPE_CHECKING:
@@ -195,7 +195,10 @@ def _collect_all_paths_known_to_pytask(session: Session) -> set[Path]:
     if session.config["config"]:
         known_paths.add(session.config["config"])
     known_paths.add(session.config["root"])
-    known_paths.add(session.config["database_filename"])
+
+    database_url = session.config["database_url"]
+    if database_url.drivername == "sqlite" and database_url.database:
+        known_paths.add(Path(database_url.database))
 
     # Add files tracked by git.
     if is_git_installed():
@@ -215,7 +218,7 @@ def _yield_paths_from_task(task: Task) -> Generator[Path, None, None]:
     """Yield all paths attached to a task."""
     yield task.path
     for attribute in ("depends_on", "produces"):
-        for node in tree_just_yield(getattr(task, attribute)):
+        for node in tree_leaves(getattr(task, attribute)):
             if hasattr(node, "path") and isinstance(node.path, Path):
                 yield node.path