+++ src/_pytask/mark/structures.py
@@ -167,6 +167,12 @@ def store_mark(obj: Callable[..., Any], mark: Mark) -> None:
         )
 
 
+_DEPRECATION_DECORATOR = """'@pytask.mark.{}' is deprecated starting pytask \
+v0.4.0 and will be removed in v0.5.0. To upgrade your project to the new syntax, read \
+the tutorial on product and dependencies: https://tinyurl.com/yrezszr4.
+"""
+
+
 class MarkGenerator:
     """Factory for :class:`MarkDecorator` objects.
 
@@ -191,16 +197,25 @@ class MarkGenerator:
         if name[0] == "_":
             raise AttributeError("Marker name must NOT start with underscore")
 
+        if name in ("depends_on", "produces"):
+            warnings.warn(
+                _DEPRECATION_DECORATOR.format(name),
+                category=DeprecationWarning,
+                stacklevel=1,
+            )
+
         # If the name is not in the set of known marks after updating,
         # then it really is time to issue a warning or an error.
         if self.config is not None and name not in self.config["markers"]:
             if self.config["strict_markers"]:
                 raise ValueError(f"Unknown pytask.mark.{name}.")
-            # Raise a specific error for common misspellings of "parametrize".
-            if name in ("parameterize", "parametrise", "parameterise"):
-                warnings.warn(
-                    f"Unknown {name!r} mark, did you mean 'parametrize'?", stacklevel=1
-                )
+
+            if name in ("parametrize", "parameterize", "parametrise", "parameterise"):
+                raise NotImplementedError(
+                    "@pytask.mark.parametrize has been removed since pytask v0.4. "
+                    "Upgrade your parametrized tasks to the new syntax defined in"
+                    "https://tinyurl.com/pytask-loops or revert to v0.3."
+                ) from None
 
             warnings.warn(
                 f"Unknown pytask.mark.{name} - is this a typo? You can register "