+++ src/_pytask/parameters.py
@@ -6,6 +6,10 @@ from pathlib import Path
 import click
 from _pytask.config import hookimpl
 from _pytask.config_utils import set_defaults_from_config
+from click import Context
+from sqlalchemy.engine import make_url
+from sqlalchemy.engine import URL
+from sqlalchemy.exc import ArgumentError
 
 
 _CONFIG_OPTION = click.Option(
@@ -67,11 +71,39 @@ _EDITOR_URL_SCHEME_OPTION = click.Option(
 """click.Option: An option to embed URLs in task ids."""
 
 
+def _database_url_callback(
+    ctx: Context, name: str, value: str | None  # noqa: ARG001
+) -> URL:
+    """Check the url for the database."""
+    # Since sqlalchemy v2.0.19, we need to shortcircuit here.
+    if value is None:
+        return None
+
+    try:
+        return make_url(value)
+    except ArgumentError:
+        raise click.BadParameter(
+            "The 'database_url' must conform to sqlalchemy's url standard: "
+            "https://docs.sqlalchemy.org/en/latest/core/engines.html"
+            "#backend-specific-urls."
+        ) from None
+
+
+_DATABASE_URL_OPTION = click.Option(
+    ["--database-url"],
+    type=str,
+    help=("Url to the database."),
+    default=None,
+    show_default="sqlite:///.../.pytask.sqlite3",
+    callback=_database_url_callback,
+)
+
+
 @hookimpl(trylast=True)
 def pytask_extend_command_line_interface(cli: click.Group) -> None:
     """Register general markers."""
     for command in ("build", "clean", "collect", "dag", "profile"):
-        cli.commands[command].params.append(_PATH_ARGUMENT)
+        cli.commands[command].params.extend([_PATH_ARGUMENT, _DATABASE_URL_OPTION])
     for command in ("build", "clean", "collect", "dag", "markers", "profile"):
         cli.commands[command].params.append(_CONFIG_OPTION)
     for command in ("build", "clean", "collect", "profile"):