+++ src/_pytask/report.py
@@ -15,7 +15,7 @@ from attrs import field
 
 
 if TYPE_CHECKING:
-    from _pytask.nodes import MetaNode
+    from _pytask.node_protocols import MetaNode
     from _pytask.nodes import Task