+++ tests/test_database.py
@@ -4,10 +4,11 @@ import textwrap
 
 import pytest
 from _pytask.database_utils import create_database
+from _pytask.database_utils import DatabaseSession
 from _pytask.database_utils import State
-from pony import orm
 from pytask import cli
 from pytask import ExitCode
+from sqlalchemy.engine import make_url
 
 
 @pytest.mark.end_to_end()
@@ -18,7 +19,7 @@ def test_existence_of_hashes_in_db(tmp_path, runner):
 
     @pytask.mark.depends_on("in.txt")
     @pytask.mark.produces("out.txt")
-    def task_write(produces):
+    def task_write(depends_on, produces):
         produces.touch()
     """
     task_path = tmp_path.joinpath("task_module.py")
@@ -30,14 +31,11 @@ def test_existence_of_hashes_in_db(tmp_path, runner):
 
     assert result.exit_code == ExitCode.OK
 
-    with orm.db_session:
-        create_database(
-            "sqlite",
-            tmp_path.joinpath(".pytask.sqlite3").as_posix(),
-            create_db=True,
-            create_tables=False,
-        )
+    create_database(
+        make_url("sqlite:///" + tmp_path.joinpath(".pytask.sqlite3").as_posix())
+    )
 
+    with DatabaseSession() as session:
         task_id = task_path.as_posix() + "::task_write"
         out_path = tmp_path.joinpath("out.txt")
 
@@ -46,26 +44,29 @@ def test_existence_of_hashes_in_db(tmp_path, runner):
             (in_path.as_posix(), in_path),
             (out_path.as_posix(), out_path),
         ):
-            modification_time = State[task_id, id_].modification_time
+            modification_time = session.get(State, (task_id, id_)).modification_time
             assert float(modification_time) == path.stat().st_mtime
 
 
 @pytest.mark.end_to_end()
 def test_rename_database_w_config(tmp_path, runner):
     """Modification dates of input and output files are stored in database."""
+    path_to_db = tmp_path.joinpath(".db.sqlite")
     tmp_path.joinpath("pyproject.toml").write_text(
-        "[tool.pytask.ini_options]\ndatabase_filename='.db.sqlite3'"
+        "[tool.pytask.ini_options]\ndatabase_url='sqlite:///.db.sqlite'"
     )
     result = runner.invoke(cli, [tmp_path.as_posix()])
     assert result.exit_code == ExitCode.OK
-    tmp_path.joinpath(".db.sqlite3").exists()
+    assert path_to_db.exists()
 
 
 @pytest.mark.end_to_end()
 def test_rename_database_w_cli(tmp_path, runner):
     """Modification dates of input and output files are stored in database."""
+    path_to_db = tmp_path.joinpath(".db.sqlite")
     result = runner.invoke(
-        cli, ["--database-filename", ".db.sqlite3", tmp_path.as_posix()]
+        cli,
+        ["--database-url", "sqlite:///.db.sqlite", tmp_path.as_posix()],
     )
     assert result.exit_code == ExitCode.OK
-    tmp_path.joinpath(".db.sqlite3").exists()
+    assert path_to_db.exists()