+++ tests/test_task.py
@@ -34,35 +34,6 @@ def test_task_with_task_decorator(tmp_path, func_name, task_name):
 
 
 @pytest.mark.end_to_end()
-@pytest.mark.parametrize("func_name", ["task_example", "func"])
-@pytest.mark.parametrize("task_name", ["the_only_task", None])
-def test_task_with_task_decorator_with_parametrize(tmp_path, func_name, task_name):
-    task_decorator_input = f"{task_name!r}" if task_name else task_name
-    source = f"""
-    import pytask
-
-    @pytask.mark.task({task_decorator_input})
-    @pytask.mark.parametrize("produces", ["out_1.txt", "out_2.txt"])
-    def {func_name}(produces):
-        produces.write_text("Hello. It's me.")
-    """
-    path_to_module = tmp_path.joinpath("task_module.py")
-    path_to_module.write_text(textwrap.dedent(source))
-
-    session = main({"paths": tmp_path})
-
-    assert session.exit_code == ExitCode.OK
-
-    file_name = path_to_module.name
-    if task_name:
-        assert session.tasks[0].name.endswith(f"{file_name}::{task_name}[out_1.txt]")
-        assert session.tasks[1].name.endswith(f"{file_name}::{task_name}[out_2.txt]")
-    else:
-        assert session.tasks[0].name.endswith(f"{file_name}::{func_name}[out_1.txt]")
-        assert session.tasks[1].name.endswith(f"{file_name}::{func_name}[out_2.txt]")
-
-
-@pytest.mark.end_to_end()
 def test_parametrization_in_for_loop(tmp_path, runner):
     source = """
     import pytask
@@ -178,7 +149,7 @@ def test_parametrization_in_for_loop_with_ids(tmp_path, runner):
     for i in range(2):
 
         @pytask.mark.task(
-            "deco_task", id=i, kwargs={"i": i, "produces": f"out_{i}.txt"}
+            "deco_task", id=str(i), kwargs={"i": i, "produces": f"out_{i}.txt"}
         )
         def example(produces, i):
             produces.write_text(str(i))
@@ -429,3 +400,158 @@ def test_that_dynamically_creates_tasks_are_captured(runner, tmp_path):
     assert "task_example[0]" in result.output
     assert "task_example[1]" in result.output
     assert "Collected 2 tasks" in result.output
+
+
+@pytest.mark.end_to_end()
+@pytest.mark.parametrize(
+    "irregular_id", [1, (1,), [1], {1}, ["a"], list("abc"), ((1,), (2,)), ({0}, {1})]
+)
+def test_raise_errors_for_irregular_ids(runner, tmp_path, irregular_id):
+    source = f"""
+    import pytask
+
+    @pytask.mark.task(id={irregular_id})
+    def task_example():
+        pass
+    """
+    tmp_path.joinpath("task_module.py").write_text(textwrap.dedent(source))
+
+    result = runner.invoke(cli, [tmp_path.as_posix()])
+
+    assert result.exit_code == ExitCode.COLLECTION_FAILED
+    assert "Argument 'id' of @pytask.mark.task" in result.output
+
+
+@pytest.mark.end_to_end()
+@pytest.mark.xfail(reason="Should fail. Mandatory products will fix the issue.")
+def test_raise_error_if_parametrization_produces_non_unique_tasks(tmp_path):
+    source = """
+    import pytask
+
+    for i in [0, 0]:
+        @pytask.mark.task(id=str(i))
+        def task_func(i=i):
+            pass
+    """
+    tmp_path.joinpath("task_module.py").write_text(textwrap.dedent(source))
+    session = main({"paths": tmp_path})
+
+    assert session.exit_code == ExitCode.COLLECTION_FAILED
+    assert isinstance(session.collection_reports[0].exc_info[1], ValueError)
+
+
+@pytest.mark.end_to_end()
+def test_task_receives_unknown_kwarg(runner, tmp_path):
+    source = """
+    import pytask
+
+    @pytask.mark.task(kwargs={"i": 1})
+    def task_example(): pass
+    """
+    tmp_path.joinpath("task_module.py").write_text(textwrap.dedent(source))
+    result = runner.invoke(cli, [tmp_path.as_posix()])
+    assert result.exit_code == ExitCode.FAILED
+
+
+@pytest.mark.end_to_end()
+def test_task_receives_namedtuple(runner, tmp_path):
+    source = """
+    import pytask
+    from typing_extensions import NamedTuple, Annotated
+    from pathlib import Path
+    from pytask import Product, PythonNode
+
+    class Args(NamedTuple):
+        path_in: Path
+        arg: str
+        path_out: Path
+
+
+    args = Args(Path("input.txt"), "world!", Path("output.txt"))
+
+    @pytask.mark.task(kwargs=args)
+    def task_example(
+        path_in: Path,
+        arg: Annotated[str, PythonNode(hash=True)],
+        path_out: Annotated[Path, Product]
+    ) -> None:
+        path_out.write_text(path_in.read_text() + " " + arg)
+    """
+    tmp_path.joinpath("task_module.py").write_text(textwrap.dedent(source))
+    tmp_path.joinpath("input.txt").write_text("Hello")
+
+    result = runner.invoke(cli, [tmp_path.as_posix()])
+    assert result.exit_code == ExitCode.OK
+    assert tmp_path.joinpath("output.txt").read_text() == "Hello world!"
+
+
+@pytest.mark.end_to_end()
+def test_task_kwargs_overwrite_default_arguments(runner, tmp_path):
+    source = """
+    import pytask
+    from pytask import Product
+    from pathlib import Path
+    from typing_extensions import Annotated
+
+    @pytask.mark.task(kwargs={
+        "in_path": Path("in.txt"), "addition": "world!", "out_path": Path("out.txt")
+    })
+    def task_example(
+        in_path: Path = Path("not_used_in.txt"),
+        addition: str = "planet!",
+        out_path: Annotated[Path, Product] = Path("not_used_out.txt"),
+    ) -> None:
+        out_path.write_text(in_path.read_text() + addition)
+    """
+    tmp_path.joinpath("task_module.py").write_text(textwrap.dedent(source))
+    tmp_path.joinpath("in.txt").write_text("Hello ")
+
+    result = runner.invoke(cli, [tmp_path.as_posix()])
+
+    assert result.exit_code == ExitCode.OK
+    assert tmp_path.joinpath("out.txt").read_text() == "Hello world!"
+    assert not tmp_path.joinpath("not_used_out.txt").exists()
+
+
+@pytest.mark.end_to_end()
+def test_return_with_task_decorator(runner, tmp_path):
+    source = """
+    from pathlib import Path
+    from typing import Any
+    from typing_extensions import Annotated
+    from pytask import PathNode
+    import pytask
+
+    node = PathNode.from_path(Path(__file__).parent.joinpath("file.txt"))
+
+    @pytask.mark.task(produces=node)
+    def task_example():
+        return "Hello, World!"
+    """
+    tmp_path.joinpath("task_module.py").write_text(textwrap.dedent(source))
+    result = runner.invoke(cli, [tmp_path.as_posix()])
+    assert result.exit_code == ExitCode.OK
+    assert tmp_path.joinpath("file.txt").read_text() == "Hello, World!"
+
+
+@pytest.mark.end_to_end()
+def test_return_with_tuple_and_task_decorator(runner, tmp_path):
+    source = """
+    from pathlib import Path
+    from typing import Any
+    from typing_extensions import Annotated
+    from pytask import PathNode
+    import pytask
+
+    node1 = PathNode.from_path(Path(__file__).parent.joinpath("file1.txt"))
+    node2 = PathNode.from_path(Path(__file__).parent.joinpath("file2.txt"))
+
+    @pytask.mark.task(produces=(node1, node2))
+    def task_example():
+        return "Hello,", "World!"
+    """
+    tmp_path.joinpath("task_module.py").write_text(textwrap.dedent(source))
+    result = runner.invoke(cli, [tmp_path.as_posix()])
+    assert result.exit_code == ExitCode.OK
+    assert tmp_path.joinpath("file1.txt").read_text() == "Hello,"
+    assert tmp_path.joinpath("file2.txt").read_text() == "World!"